+++ tests/layout/test_layout_control_mode.py
@@ -20,7 +20,7 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import pytest
-
+import asyncio
 import logging
 
 from PySide2 import QtCore
@@ -49,31 +49,45 @@ def widget(qtbot):
     return widget
 
 
-def test_callback_signal_control_normal(qtbot, widget):
+async def test_callback_signal_control_normal(qtbot, widget):
     widget.layout_control_mode.model.report_control_status()
+
+    # Sleep so the event loop can access CPU to handle the signal
+    await asyncio.sleep(1)
 
     assert widget.layout_control_mode._button_open_loop.isEnabled() is False
     assert widget.layout_control_mode._button_closed_loop.isEnabled() is False
 
 
-def test_callback_signal_control_prohibit_control(qtbot, widget):
+async def test_callback_signal_control_prohibit_control(qtbot, widget):
     widget.layout_control_mode.model.local_mode = LocalMode.Enable
 
     widget.layout_control_mode.model.report_control_status()
+
+    # Sleep so the event loop can access CPU to handle the signal
+    await asyncio.sleep(1)
 
     assert widget.layout_control_mode._button_open_loop.isEnabled() is False
     assert widget.layout_control_mode._button_closed_loop.isEnabled() is True
 
 
-def test_switch_force_balance_system(qtbot, widget):
+async def test_switch_force_balance_system(qtbot, widget):
     widget.layout_control_mode.model.local_mode = LocalMode.Enable
 
     widget.layout_control_mode.switch_force_balance_system(True)
+
+    # Sleep so the event loop can access CPU to handle the signal
+    await asyncio.sleep(1)
+
     assert widget.layout_control_mode.model.is_closed_loop is True
     assert widget.layout_control_mode._button_open_loop.isEnabled() is True
     assert widget.layout_control_mode._button_closed_loop.isEnabled() is False
 
     qtbot.mouseClick(widget.layout_control_mode._button_open_loop, QtCore.Qt.LeftButton)
+
+    # Sleep so the event loop can access CPU to handle the signal
+    await asyncio.sleep(1)
+
     assert widget.layout_control_mode.model.is_closed_loop is False
     assert widget.layout_control_mode._button_open_loop.isEnabled() is False
     assert widget.layout_control_mode._button_closed_loop.isEnabled() is True
@@ -87,6 +101,10 @@ def test_switch_force_balance_system(qtbot, widget):
     qtbot.mouseClick(
         widget.layout_control_mode._button_closed_loop, QtCore.Qt.LeftButton
     )
+
+    # Sleep so the event loop can access CPU to handle the signal
+    await asyncio.sleep(1)
+
     assert widget.layout_control_mode.model.is_closed_loop is True
     assert widget.layout_control_mode._button_open_loop.isEnabled() is True
     assert widget.layout_control_mode._button_closed_loop.isEnabled() is False